split main.js to multiple files
